requirements.txt
# Pillow==8.4.0
# PyInstaller==3.6
#pytube==12.0.0
#requests==2.27.1
#urllib3==1.26.7
Pillow==11.0.0
PyInstaller==6.11.1
pytube==15.0.0
yt_dlp==2025.1.26
requests==2.32.3
urllib3==2.2.3

